VisionModel/main.py

from groq import Groq
import os
from dotenv import load_dotenv
load_dotenv()

client = Groq(api_key=os.environ.get("GROQ_API_KEY"))
completion = client.chat.completions.create(
    model="meta-llama/llama-4-scout-17b-16e-instruct",
    messages=[
        {
            "role": "user",
            "content": [
                {
                    "type": "text",
                    "text": "What's this image?"
                },
                {
                    "type": "image_url",
                    "image_url": {
                        "url": "https://m4b6f3p8.delivery.rocketcdn.me/app/uploads/2021/04/tineaCorporisRingwormofBody_31362_lg.jpg"
                    }
                }
            ]
        }
    ],
    temperature=1,
    max_completion_tokens=1024,
    top_p=1,
    stream=False,
    stop=None,
)

print(completion.choices[0].message)
